# backend/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: BRUSKA REST API
  version: 1.0.0
  description: |
    Документация REST API проекта BRUSKA.
    Все маршруты имеют префикс `/bruska`.
servers:
  - url: /bruska
...
paths:
...
  /admin/products:
    get:
      tags: [admin-products]
      summary: Получить список товаров (админ)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogItem'
    post:
      tags: [admin-products]
      summary: Создать товар
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDraft'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItem'
        '400':
          description: Ошибка валидации или некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/products/{id}:
    get:
      tags: [admin-products]
      summary: Получить товар по ID (админ)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Товар найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItem'
        '400':
          description: Некорректный ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [admin-products]
      summary: Удалить товар
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Товар удален
        '400':
          description: Некорректный ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CatalogColor:
      type: object
      properties:
        tag:
          type: string
      required: [tag]
    CatalogSize:
      type: object
      properties:
        length:
          type: integer
        width:
          type: integer
        height:
          type: integer
      required: [length, width, height]
    CatalogItem:
      type: object
      properties:
        id:
          type: integer
        tag:
          type: string
        title:
          type: string
        dir:
          type: string
        img:
          type: string
        sliderImages:
          type: array
          items:
            type: string
        price:
          type: object
          additionalProperties:
            type: number
        priceTag:
          type: string
        withBtn:
          type: boolean
        colors:
          type: array
          items:
            $ref: '#/components/schemas/CatalogColor'
        sizes:
          type: array
          items:
            $ref: '#/components/schemas/CatalogSize'
        description:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        onPallet:
          type: number
        weight:
          type: number
      required:
        - id
        - title
        - dir
        - img
        - sliderImages
        - price
        - priceTag
        - colors
        - sizes
        - description
        - isAvailable
        - onPallet
        - weight
    ProductColorDraft:
      type: object
      properties:
        tag:
          type: string
        price:
          type: number
      required: [tag, price]
    ProductDraft:
      type: object
      properties:
        category:
          type: string
        title:
          type: string
        dir:
          type: string
        img:
          type: string
        priceTag:
          type: string
        basePrice:
          type: number
        color:
          type: string
        colors:
          type: array
          items:
            $ref: '#/components/schemas/ProductColorDraft'
        description:
          type: string
        onPallet:
          type: number
        weight:
          type: number
        length:
          type: integer
        width:
          type: integer
        height:
          type: integer
        withBtn:
          type: boolean
        isAvailable:
          type: boolean
      required:
        - category
        - title
        - dir
        - img
        - priceTag
        - colors
        - onPallet
        - weight
        - length
        - width
        - height
        - withBtn
        - isAvailable
...